[FIX] Replaced int by size_t

+++ src/my_strlen.c
@@ -5,9 +5,11 @@
 ** strlen rewritten :)
 */
 
-int my_strlen(char const *str)
+#include "my_strings.h"
+
+size_t my_strlen(char const *str)
 {
-    int size = 0;
+    size_t size = 0;
 
     while (*str++ != '\0')
         ++size;
